temporarily disable gate (#76)
the mmu pipeline is failing on the indexer step. After running
the API call the indexer does, we see gate is one of the markets
returned and has a valid status value, so the failures in the mmu
pipeline currently are difficult to explain. As we investigate,
we can disable just gate while we sort out the root cause of the
problem.